% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughViz.R
\name{renderRoughViz}
\alias{renderRoughViz}
\title{Shiny render function for roughViz}
\usage{
renderRoughViz(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression to render}

\item{env}{Environment}

\item{quoted}{Is expression quoted}
}
\description{
Shiny render function for roughViz
}
