% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughViz.R
\name{roughScatter}
\alias{roughScatter}
\title{Create a rough scatter plot}
\usage{
roughScatter(
  data,
  x = NULL,
  y = NULL,
  title = NULL,
  width = NULL,
  height = NULL,
  colorVar = NULL,
  colors = NULL,
  highlightLabel = NULL,
  roughness = 1,
  fillStyle = "hachure",
  fillWeight = 0.5,
  strokeWidth = 1,
  bowing = 0,
  simplification = 0.2,
  font = 0,
  margin = NULL,
  axisFontSize = "1rem",
  axisRoughness = 0.5,
  axisStrokeWidth = 0.5,
  radius = 8,
  innerStrokeWidth = 1,
  titleAlign = "center",
  ...
)
}
\arguments{
\item{data}{Data frame or list containing the data}

\item{x}{Column name for x values}

\item{y}{Column name for y values}

\item{title}{Chart title}

\item{width}{Chart width}

\item{height}{Chart height}

\item{colorVar}{Column name for color variable}

\item{colors}{Array of colors}

\item{highlightLabel}{Label to highlight}

\item{roughness}{Roughness level}

\item{fillStyle}{Fill style}

\item{fillWeight}{Fill weight}

\item{strokeWidth}{Stroke width}

\item{bowing}{Bowing effect}

\item{simplification}{Simplification level}

\item{font}{Font style}

\item{margin}{List with margins}

\item{axisFontSize}{Axis font size}

\item{axisRoughness}{Axis roughness}

\item{axisStrokeWidth}{Axis stroke width}

\item{radius}{Circle radius}

\item{innerStrokeWidth}{Inner stroke width}

\item{...}{Additional configuration options}
}
\description{
Create a rough scatter plot
}
