% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughViz.R
\name{roughStackedBar}
\alias{roughStackedBar}
\title{Create a rough stacked bar chart}
\usage{
roughStackedBar(
  data,
  labels = NULL,
  title = NULL,
  width = NULL,
  height = NULL,
  colors = NULL,
  roughness = 1,
  fillStyle = "hachure",
  fillWeight = NULL,
  strokeWidth = 1,
  bowing = 0,
  simplification = 0.2,
  font = 0,
  margin = list(top = 50, right = 20, bottom = 70, left = 150),
  axisFontSize = "1rem",
  axisRoughness = 0.5,
  axisStrokeWidth = 0.5,
  innerStrokeWidth = 0,
  padding = 0.1,
  titleAlign = "center",
  ...
)
}
\arguments{
\item{data}{Data frame or list containing the data}

\item{labels}{Column name for labels}

\item{title}{Chart title}

\item{width}{Chart width}

\item{height}{Chart height}

\item{colors}{Array of colors}

\item{roughness}{Roughness level}

\item{fillStyle}{Fill style}

\item{fillWeight}{Fill weight}

\item{strokeWidth}{Stroke width}

\item{bowing}{Bowing effect}

\item{simplification}{Simplification level}

\item{font}{Font style}

\item{margin}{List with margins}

\item{axisFontSize}{Axis font size}

\item{axisRoughness}{Axis roughness}

\item{axisStrokeWidth}{Axis stroke width}

\item{innerStrokeWidth}{Inner stroke width}

\item{padding}{Padding between bars}

\item{...}{Additional configuration options}
}
\description{
Create a rough stacked bar chart
}
