% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughViz.R
\name{roughViz}
\alias{roughViz}
\title{Create a roughViz widget}
\usage{
roughViz(chartType, config, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{chartType}{The type of chart to create}

\item{config}{Configuration list for the chart}

\item{width}{Widget width}

\item{height}{Widget height}

\item{elementId}{Element ID}
}
\description{
Create a roughViz widget
}
