% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughViz.R
\name{roughVizOutput}
\alias{roughVizOutput}
\title{Shiny output for roughViz}
\usage{
roughVizOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output ID}

\item{width}{Width}

\item{height}{Height}
}
\description{
Shiny output for roughViz
}
